Conditional Breakpoints
When you select “Edit Breakpoint” you are presented with a context
menu showing additional options for that breakpoint. You can set  a
condition that will cause your breakpoint to trigger only when that
condition is met, you can cause it to ignore the breakpoint until a
certain number of times, and can even trigger complex actions to be
performed automatically when that breakpoint is hit.